pyproject.toml
[project]
name = "ogcapi-registry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Claude", email = "noreply@anthropic.com" },
    { name = "Francesco Bartoli", email = "xbartolone@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "openapi-pydantic>=0.5.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "pytest-httpx>=0.36.0",
    "ruff>=0.14.8",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
plugins = ["pydantic.mypy"]
# Ignore missing imports for packages without type stubs
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
